initial Editor class
