+++ package.json
@@ -17,7 +17,8 @@
     "js-yaml": "^3.13.1",
     "material-design-icons": "^3.0.1",
     "path": "^0.12.7",
-    "python-shell": "^1.0.8"
+    "python-shell": "^1.0.8",
+    "walk": "^2.3.14"
   },
   "devDependencies": {
     "electron": "^6.0.10"
