+++ renderer/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="./assets/workbench_desktop.css">
+    <link rel="stylesheet" href="../assets/workbench_desktop.css">
     <title>Islandora Workbench Desktop</title>
   </head>
   <body>
@@ -12,6 +12,6 @@
     <p id="workbench-config-file"></p>
     <p id="workbench-output"></p>
     <p id="workbench-exit"></p>
-    <script src="./renderer.js"></script>
+    <script src="./index.js"></script>
   </body>
 </html>
