+++ renderer/index.js
@@ -1,6 +1,4 @@
-// This file is required by the index.html file and will
-// be executed in the renderer process for that window.
-// All of the Node.js APIs are available in this process.
+'use strict'
 
 const ipc = require('electron').ipcRenderer
 const asyncRunButton = document.getElementById('runButton')
